optimistic locking - concurrencyVersion concept change

+++ clp/src/main/java/pl/biltec/yaess/clp/domain/customer/exception/ConcurrentModificationException.java
@@ -5,7 +5,7 @@ import pl.biltec.yaess.clp.domain.customer.CustomerId;
 
 public class ConcurrentModificationException extends DomainOperationException {
 
-	public ConcurrentModificationException(CustomerId customerId, int expectedConcurrencyVersion, Integer actualConcurrencyVersion) {
+	public ConcurrentModificationException(CustomerId customerId, long expectedConcurrencyVersion, long actualConcurrencyVersion) {
 
 		super(customerId, String.format("Expected Concurrency Version: %s, found %s", expectedConcurrencyVersion, actualConcurrencyVersion));
 	}
